stubbed out copy_pw

+++ interface.py
@@ -45,6 +45,29 @@ def retrieve_info_menu():
     print('Retrieve an Account')
     results = lookup_choice_menu()
 
+    print('Results: ')
+    for i in range(0, len(results)):
+        print(i, ': ', impl.acct_directory[results[i]])
+    
+    if len(results) > 0:
+        right_choice = False
+        while not right_choice:
+            account_to_retrieve = input('Type the number of the account whose password you wish to copy: ')
+            confirm_retrieval = input('Are you sure that you want to retrieve the password of account number ' + account_to_retrieve + '? (y or n) ')
+            if confirm_retrieval =='y':
+                #Currently doesn't deal with cases where the user doesn't type a number
+                if int(account_to_retrieve) < len(results):
+                    right_choice = True
+                    index_to_retrieve = results[int(account_to_retrieve)]
+                    impl.copy_pw(index_to_retrieve)
+                else:
+                    print('That index is not valid. Please enter a valid index')
+            elif confirm_retrieval == 'n':
+                right_choice = True
+                print('Canceling retrieval...')
+                #Return to the results 
+    else:
+        print('No accounts matched your search')
 
 def register_menu():
     print('Register a new account')
@@ -77,22 +100,24 @@ def delete_acct_menu():
     for i in range(0, len(results)):
         print(i, ': ', impl.acct_directory[results[i]])
 
-    right_choice = False
-    while not right_choice:
-    	account_to_delete = input('Type the number of the account you wish to delete: ')
-    	confirm_delete = input('Are you sure that you want to delete account number ' + account_to_delete + '? (y or n) ')
-    	if confirm_delete =='y':
-    		#Currently doesn't deal with cases where the user doesn't type a number
-    		if int(account_to_delete) < len(results):
-    			right_choice = True
-    			index_to_delete = results[int(account_to_delete)]
-    			impl.delete_acct(index_to_delete)
-    		else:
-    			print('That index is not valid. Please enter a valid index')
-    	elif confirm_delete == 'n':
-    		right_choice = True
-    		print('Canceling delete...')
-    		#Return to the results 
+    if len(results) > 0:
+        right_choice = False
+        while not right_choice:
+            account_to_delete = input('Type the number of the account you wish to delete: ')
+            confirm_delete = input('Are you sure that you want to delete account number ' + account_to_delete + '? (y or n) ')
+            if confirm_delete =='y':
+                #Currently doesn't deal with cases where the user doesn't type a number
+                if int(account_to_delete) < len(results):
+                    right_choice = True
+                    index_to_delete = results[int(account_to_delete)]
+                    impl.delete_acct(index_to_delete)
+                else:
+                    print('That index is not valid. Please enter a valid index')
+            elif confirm_delete == 'n':
+                right_choice = True
+                print('Canceling delete...')
+    else:
+        print('No accounts matched your search')
 
 
 def change_master_pw_menu():
@@ -113,23 +138,23 @@ def setup_menu():
 
 
 def lookup_choice_menu():
-	results = []
-	is_valid_choice = False
-	while not is_valid_choice:
-		lookup_choice = input('Look up account by service name (name), service URL (url), or username (user)?')
-		if lookup_choice == 'name':
-			is_valid_choice = True
-			account_name = input('Service Name: ')
-			results = impl.search_by_service_name(account_name)
-		elif lookup_choice == 'url':
-			is_valid_choice = True
-			account_url = input('Service URL: ')
-			results = impl.search_by_url(account_url)
-		elif lookup_choice == 'user':
-			is_valid_choice = True
-			account_username = input('Username: ')
-			results = impl.search_by_username(account_username)
-	return results
+    results = []
+    is_valid_choice = False
+    while not is_valid_choice:
+        lookup_choice = input('Look up account by service name (name), service URL (url), or username (user)?')
+        if lookup_choice == 'name':
+            is_valid_choice = True
+            account_name = input('Service Name: ')
+            results = impl.search_by_service_name(account_name)
+        elif lookup_choice == 'url':
+            is_valid_choice = True
+            account_url = input('Service URL: ')
+            results = impl.search_by_url(account_url)
+        elif lookup_choice == 'user':
+            is_valid_choice = True
+            account_username = input('Username: ')
+            results = impl.search_by_username(account_username)
+    return results
 
 
 def notify_login_expired():
